Included a "Reader" in py2store.base. Doing (ReadOnlyMixin, Persister) every time I need a "reader" seems way to funky, and even calling a class a persister when really it just reads seems misleading. I'll try out the "Reader" approach for a bit and see if it sticks.
Changed the names of the two "persisters" (now "readers") to reflect the "reader" term, but also that this particular reader makes a full local copy before giving access, and also removed "Shared" because doesn't add enough information to the already crammed name.

DropboxFolderCopyReader: Makes a full local copy of the folder (by default, to a local temp folder) and gives access to it. This can be useful in some situations, but in that case, I think it should be very clear that this is what's happening, and though the dropbox reader should be in the dropbox module, it should be nothing else than a local folder files reader with a dropbox-specific init. Because one the data is copied, it's a local store.

This automatic full copy can be sometimes undesirable (for example, perhaps the user just wants a few files, or even none, but still would like to explore the contents of the remote folder). There is a need for a reader that just performs the operations, as they're requested. It should list the files available when `__iter__` is called, it should provide the contents of a file when `__getitem__` is called, etc.

Of course, it's probably a good idea to cache the file (name/path) list once local has it, and might even be good to cache (in a persisted local file), the file contents as well. But these are cache concerns, so need to be decoupled. That is, even if we cache by default, the caching mechanism should be decoupled, and probably exposed as a default argument of the init, that the use can override if he so chooses.

Similar comments for DropboxFileCopyReader.